chore: moved log-entry files to indexer-logger
