+++ runner/src/indexer-logger/log-entry.test.ts
